Bug 38446: BACnet MSTP UI - Observations

Related work items: #38446

+++ logic/src/main/java/a75f/io/logic/bo/building/system/BacnetServicesUtils.kt
@@ -183,7 +183,7 @@ class BacnetServicesUtils: BacnetRequestUtil {
         ) {
             selectedValue
         } else {
-            (selectedValue.toInt() + 1).toString()
+            (selectedValue.toDouble().toInt() + 1).toString()
         }
 
         val objectIdentifierBacNet = ObjectIdentifierBacNet(objectTypeEnum.value, objectId.toString())
